Replaced table types
