+++ src/components/assets-datatable.tsx
@@ -10,11 +10,72 @@ import {
 	TableHeader,
 	TableRow
 } from '@/components/ui/table';
+import {
+	DropdownMenu,
+	DropdownMenuContent,
+	DropdownMenuItem,
+	DropdownMenuLabel,
+	DropdownMenuTrigger
+} from './ui/dropdown-menu';
+import { Button } from './ui/button';
+import { MoreHorizontal } from 'lucide-react';
+import { Asset } from '@/lib/types';
 
 interface DataTableProps<TData, TValue> {
 	columns: ColumnDef<TData, TValue>[];
 	data: TData[];
 }
+
+export type VaultType = {
+	asset_type: string;
+	id: string;
+	symbol: string;
+	amount: string;
+};
+export const columns: ColumnDef<Asset>[] = [
+	// {
+	// 	accessorKey: 'asset_type',
+	// 	header: 'Asset type'
+	// },
+	{
+		accessorKey: 'faucetId',
+		header: 'Faucet ID'
+	},
+	// {
+	// 	accessorKey: 'symbol',
+	// 	header: 'Symbol'
+	// },
+	{
+		accessorKey: 'amount',
+		header: 'Amount'
+	},
+	{
+		id: 'actions',
+		enableHiding: false,
+		cell: ({ row }) => {
+			const faucet = row.original;
+
+			return (
+				<DropdownMenu>
+					<DropdownMenuTrigger asChild>
+						<Button className="h-8 w-8 p-0 hover:bg-theme-border">
+							<span className="sr-only">Open menu</span>
+							<MoreHorizontal />
+						</Button>
+					</DropdownMenuTrigger>
+					<DropdownMenuContent align="end">
+						<DropdownMenuLabel>Actions</DropdownMenuLabel>
+						<DropdownMenuItem
+							onClick={() => navigator.clipboard.writeText(faucet.faucetId.toString())}
+						>
+							Copy faucet ID
+						</DropdownMenuItem>
+					</DropdownMenuContent>
+				</DropdownMenu>
+			);
+		}
+	}
+];
 
 export function AssetsDatatable<TData, TValue>({ columns, data }: DataTableProps<TData, TValue>) {
 	const table = useReactTable({
@@ -22,7 +83,7 @@ export function AssetsDatatable<TData, TValue>({ columns, data }: DataTableProps
 		columns,
 		getCoreRowModel: getCoreRowModel()
 	});
-
+	console.log(data);
 	return (
 		<div className="rounded-md border-2 border-theme-border">
 			<Table className="[&_tr:hover]:bg-transparent">
